pyproject.toml
[tool.poetry]
name = "2-pract"
version = "0.1.0"
description = ""
authors = ["Corner324 <mosenn@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"
pandas = "^2.2.3"
xlrd = "^2.0.1"
openpyxl = "^3.1.5"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
selenium = "^4.32.0"
webdriver-manager = "^4.0.2"
pydantic = "^2.11.4"
asyncpg = "^0.30.0"
aiohttp = "^3.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
